refractored - modularized config; used self
